# util/uniq: use generics instead of reflect