@andrew-d 2022-08-29
Actual benchmarks:

<details>
  <summary>Click to expand</summary>

```
name              old time/op    new time/op     delta
/AllDups/1-10        721ns ± 0%      183ns ± 2%   -74.63%  (p=0.016 n=4+5)
/AllDups/16-10       724ns ± 0%      188ns ± 5%   -74.11%  (p=0.008 n=5+5)
/AllDups/256-10      733ns ± 1%      186ns ± 5%   -74.62%  (p=0.008 n=5+5)
/AllDups/4096-10     730ns ± 1%      187ns ± 5%   -74.35%  (p=0.008 n=5+5)
/NoDups/1-10         691ns ± 1%      180ns ± 2%   -74.00%  (p=0.008 n=5+5)
/NoDups/16-10        688ns ± 1%      181ns ± 2%   -73.70%  (p=0.008 n=5+5)
/NoDups/256-10       689ns ± 1%      179ns ± 2%   -74.01%  (p=0.008 n=5+5)
/NoDups/4096-10      687ns ± 1%      181ns ± 3%   -73.66%  (p=0.008 n=5+5)

name              old speed      new speed       delta
/AllDups/1-10     88.7MB/s ± 0%  349.9MB/s ± 2%  +294.62%  (p=0.008 n=5+5)
/AllDups/16-10    88.4MB/s ± 0%  341.6MB/s ± 5%  +286.59%  (p=0.008 n=5+5)
/AllDups/256-10   87.3MB/s ± 1%  344.1MB/s ± 5%  +294.35%  (p=0.008 n=5+5)
/AllDups/4096-10  87.7MB/s ± 1%  342.0MB/s ± 4%  +290.19%  (p=0.008 n=5+5)
/NoDups/1-10      92.6MB/s ± 1%  356.2MB/s ± 2%  +284.66%  (p=0.008 n=5+5)
/NoDups/16-10     93.1MB/s ± 1%  353.9MB/s ± 2%  +280.28%  (p=0.008 n=5+5)
/NoDups/256-10    92.9MB/s ± 1%  357.5MB/s ± 2%  +284.84%  (p=0.008 n=5+5)
/NoDups/4096-10   93.1MB/s ± 1%  353.7MB/s ± 3%  +279.76%  (p=0.008 n=5+5)

name              old alloc/op   new alloc/op    delta
/AllDups/1-10        0.00B           0.00B           ~     (all equal)
/AllDups/16-10       0.00B           0.00B           ~     (all equal)
/AllDups/256-10      0.00B           0.00B           ~     (all equal)
/AllDups/4096-10     0.00B           0.00B           ~     (all equal)
/NoDups/1-10         0.00B           0.00B           ~     (all equal)
/NoDups/16-10        0.00B           0.00B           ~     (all equal)
/NoDups/256-10       0.00B           0.00B           ~     (all equal)
/NoDups/4096-10      0.00B           0.00B           ~     (all equal)

name              old allocs/op  new allocs/op   delta
/AllDups/1-10         0.00            0.00           ~     (all equal)
/AllDups/16-10        0.00            0.00           ~     (all equal)
/AllDups/256-10       0.00            0.00           ~     (all equal)
/AllDups/4096-10      0.00            0.00           ~     (all equal)
/NoDups/1-10          0.00            0.00           ~     (all equal)
/NoDups/16-10         0.00            0.00           ~     (all equal)
/NoDups/256-10        0.00            0.00           ~     (all equal)
/NoDups/4096-10       0.00            0.00           ~     (all equal)
```

</details>